Bug Fixes Galore: Current Area Checking; Fix Mouse Spazz; Optionally Use Special Attacks. Hopefully fixed: Eating in multi-combat and prayer potions.....

+++ src/com/zalgoproductions/strategies/script/potions/prayer/PrayerPotionStrategy.java
@@ -12,8 +12,9 @@ public class PrayerPotionStrategy extends PotionStrategy{
 
 	@Override
 	protected boolean needsPotion() {
-		double restoration = 70 + Math.floor(2.5 * Skills.getRealLevel(Skills.PRAYER));
+		int restoration = 70 + (int)Math.floor(2.5 * Skills.getRealLevel(Skills.PRAYER));
 		int prayerUsed = Prayer.getMaxPrayerPoints() - Prayer.getPrayerPoints();
 		return restoration <= prayerUsed;
 	}
 }
+
