+++ src/com/zalgoproductions/strategies/script/potions/prayer/SuperPrayerStrategy.java
@@ -12,7 +12,7 @@ public class SuperPrayerStrategy extends PotionStrategy{
 
 	@Override
 	protected boolean needsPotion() {
-		double restoration = 70 + Math.floor(3.43 * Skills.getRealLevel(Skills.PRAYER));
+		int restoration = 70 + (int)Math.floor(3.43 * Skills.getRealLevel(Skills.PRAYER));
 		int prayerUsed = Prayer.getMaxPrayerPoints() - Prayer.getPrayerPoints();
 		return restoration <= prayerUsed;
 	}
